feat: export public api
Define the api shape, moving also the index.ts into src/lib root
