--- src/index.ts
@@ -1,31 +1,0 @@
-import {
-  DOSource,
-  CARTOSource,
-  basicStyle,
-  colorBinsStyle,
-  colorCategoriesStyle,
-  colorContinuousStyle,
-  sizeBinsStyle,
-  sizeCategoriesStyle,
-  sizeContinuousStyle,
-  createGoogleMap,
-  createMap,
-  Layer,
-  Popup
-} from './lib/viz';
-
-export const viz = {
-  DOSource,
-  CARTOSource,
-  basicStyle,
-  colorBinsStyle,
-  colorCategoriesStyle,
-  colorContinuousStyle,
-  sizeBinsStyle,
-  sizeCategoriesStyle,
-  sizeContinuousStyle,
-  createGoogleMap,
-  createMap,
-  Layer,
-  Popup
-};
